runny: refactor to fix race condition in Command
The PTY that we create was getting shared to multiple objects.

This was causing instances of double-close(), which became a problem
later on when we went to open() a new pty.

The solution is to dup() the file descriptors before passing them
to Command, and before using them for reading or writing.

Signed-off-by: Sean Cross <sean@xobs.io>